Moved Cohen's folder and added corrs.py
corrs.py generates a "compartible" corrMatrix of the relevent dataframes to Covid data. Take note of the imputations and date indexing
