+++ data/COVID_DATA_ANALYSIS/SIR_experimentation.py
@@ -1,0 +1,83 @@
+import numpy as np
+import pandas as pd
+from scipy.integrate import solve_ivp
+from scipy.optimize import minimize
+import matplotlib.pyplot as plt
+from datetime import timedelta, datetime
+
+START_DATE = {
+  'Japan': '1/22/20',
+  'Italy': '1/31/20',
+  'Republic of Korea': '1/22/20',
+  'Iran (Islamic Republic of)': '2/19/20'
+}
+
+def main():
+	with open(fn, "rb") as f:
+		df_list = pickle.load(f)
+		df = df_list[11]
+
+	df.head()
+	df.describe()
+
+class Learner(object):
+    def __init__(self, country, loss):
+        self.country = country
+        self.loss = loss
+
+    def load_confirmed(self, country):
+      """
+      Load confirmed cases downloaded from HDX
+      """
+      df = pd.read_csv('data/time_series_19-covid-Confirmed.csv')
+      country_df = df[df['Country/Region'] == country]
+      return country_df.iloc[0].loc[START_DATE[country]:]
+
+    def extend_index(self, index, new_size):
+        values = index.values
+        current = datetime.strptime(index[-1], '%m/%d/%y')
+        while len(values) < new_size:
+            current = current + timedelta(days=1)
+            values = np.append(values, datetime.strftime(current, '%m/%d/%y'))
+        return values
+
+    def predict(self, beta, gamma, data):
+        """
+        Predict how the number of people in each compartment can be changed through time toward the future.
+        The model is formulated with the given beta and gamma.
+        """
+        predict_range = 150
+        new_index = self.extend_index(data.index, predict_range)
+        size = len(new_index)
+        def SIR(t, y):
+            S = y[0]
+            I = y[1]
+            R = y[2]
+            return [-beta*S*I, beta*S*I-gamma*I, gamma*I]
+        extended_actual = np.concatenate((data.values, [None] * (size - len(data.values))))
+        return new_index, extended_actual, solve_ivp(SIR, [0, size], [S_0,I_0,R_0], t_eval=np.arange(0, size, 1))
+
+    def train(self):
+        """
+        Run the optimization to estimate the beta and gamma fitting the given confirmed cases.
+        """
+        data = self.load_confirmed(self.country)
+        optimal = minimize(
+            loss,
+            [0.001, 0.001],
+            args=(data),
+            method='L-BFGS-B',
+            bounds=[(0.00000001, 0.4), (0.00000001, 0.4)]
+        )
+        beta, gamma = optimal.x
+        new_index, extended_actual, prediction = self.predict(beta, gamma, data)
+        df = pd.DataFrame({
+            'Actual': extended_actual,
+            'S': prediction.y[0],
+            'I': prediction.y[1],
+            'R': prediction.y[2]
+        }, index=new_index)
+        fig, ax = plt.subplots(figsize=(15, 10))
+        ax.set_title(self.country)
+        df.plot(ax=ax)
+        fig.savefig(f"{self.country}.png")
