+++ data/COVID_DATA_ANALYSIS/plots_pngs/deaths_to_cases_bar.png
