+++ data/COVID_DATA_ANALYSIS/plots_pngs/stacked_bar_deaths.png
